pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "copaw-enterprise-skills"
version = "1.0.0"
description = "Multi-skill enterprise platform built on CoPaw"
requires-python = ">=3.11"
dependencies = [
    # CoPaw framework
    "copaw>=1.0.0",

    # LLM clients
    "openai>=1.30.0",
    "google-genai>=1.0.0",

    # Data ingestion
    "feedparser>=6.0.0",
    "trafilatura>=1.6.0",
    "duckduckgo-search>=6.0.0",
    "httpx>=0.27.0",

    # Structured output
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "json-repair>=0.28.0",

    # Async
    "aiofiles>=23.0.0",

    # PPTX
    "python-pptx>=0.6.23",

    # API
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-socketio>=5.11.0",
    "python-jose[cryptography]>=3.3.0",

    # Email
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["enterprise_skills_lib*", "skills*", "api*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require network/LLM access",
]

[tool.ruff]
line-length = 120
target-version = "py311"
